src/app: use img-thumbnail in thumbnail component
For some reason, when I override the thumbnail component in my own
theme it only works for the simple item page. For now, I will just
override this in the base component so it also works in the commu-
nity and collection browse pages, and recent submission lists.

Note that I will have to resolve conflicts with the base app when
it gets updated. Hopefully there is a proper fix by then so I can
have a consistent thumbnail style using Bootstrap's img-thumbnail
class.