src: add a custom ORCID identifier component
Add a custom ORCID identifier component for the simple item page.
For now this is implemented directly in the main src/app because
we can't do it as a themed component yet.

As for the implementation, I don't know Angular at all so this is
probably a very naive implementation that I will have to re-visit
later.