src/themes/cgspace: add themed file section component
Add a themed file section component that displays the bitstream's
description in lieu of the name if it is present. As a compromise
between DSpace 7's default and the old XMLUI behavior, I also add
the file's extension as a cue to the user (via a helper function).

If the bitstream does not have a description it uses the same dis-
play as the default file selection component.