src: add item-page component for search links
This is a custom component that takes a metadata value and builds
a Discovery search link. This is more like the XMLUI search links
we had in DSpace 6. The closest thing DSpace 7.x has is the browse
by functionality, which is dependent on hierarchical vocabularies
and has too many side effects in my opinion.